Use a named group for the emoji_and_char_name regex pattern
The regex_finditer function we use from kitty only reports matches in
the last group if groups are used, but no named groups are used. The
pattern from the emoji package has a group around the emojis, so we have
to put a named group around our merged pattern to be able to match
everything.