Add typings
Unfortunately, I had to use Any a couple of places here. args in mark
should really be HintsCLIOptions, but that's typed as a class without
any attributes, so using that gives type errors. And answer in
handle_result doesn't have a type from kitty, so I'd have to create it
myself. I didn't bother because it's a bit cumbersome, and part of the
type is dependent on what's returned from mark.